Add make release-tagged, the true release path
This target builds a nomos image and a config-sync-manifest.yaml that
are to be considered released.  They correspond to a tag that is
maintained in git.

The nomos image is built with VERSION set to:
some SEMVER_TAG, like vMAJOR.MINOR.PATCH-rc.X

and pushed to:
gcr.io/config-management-release/nomos:[SEMVER_TAG]

This is spliced into config-sync-manifest.yaml, which is uploaded at:
gs://config-management-release/config-sync-manifest/tag/[SEMVER_TAG]/config-sync-manifest.yaml

See demo output in this gPaste: go/paste/4952769653899264

See go/cs-manifest-creation-and-release for more info.

Bug: 160258652
Change-Id: I83129a321720cb6375e86e581a9ce461bcf9f7f8