Fix nomos-stresstest.
The recent migration of testing to Go changed functions in
e2e/lib/setup.bash that the stresstests dependended on.  Since this was
only really the timing stuff in what was setup::common and
setup::common_teardown, replicating what those functions did in the
stresstest-specific setup() and teardown() functions is the easiest fix.
Obviously if we decide to extend stresstests someday we will want to
refactor this again.

Bug: 160325953
Change-Id: I91a866d771797337141201c084befe1902a108dc