Don't vet YAML in examples/
While the YAML in examples/ is used in tests, we aren't using FileParser
on it except in cases where FileParser behavior is under test.

Arguably we should have a better abstraction so we aren't mixing
documentation YAML and test data YAML, but that isn't worth fixing now.

The new Go e2e tests don't have the testdata/ problem as the tests
construct their own objects/YAML, so this will gradually become moot.

Bug: 160412476
Change-Id: Ic5c21e4ce77d07572341782891ee1e3b95342da8