Scale up Go tests run in Prow
Just running a single test worked fine, so scale up to running multiple
tests. Now the target runs five tests, hopefully concurrently.

It's ok that this isn't a strict superset of the CL before - we're in
debugging mode, not relying on this right now.

This will tell us two things:

1. Do bats tests execute successfully?
2. Do tests run in parallel, or sequentially?

If they run in parallel as desired, runtime should be about 150s.

Change-Id: Ic150d7595bdac2c332da37fd858fa2a999813ef2