Increase Go test timeout to 240m
Yes, this is a long time and not practical.

For now, we want to verify two things:

1. All tests can actually pass when run in sequence. This elminates many
classes of bugs in our configuration that cause individual tests to
misbehave. If this passes, we'll know previous failures were concurrency
issues. If this fails, we'll see the exact tests experiencing/causing
the issues since all tests will have had time to run in isolation.

2. The tests are not flaky. If this succeeds often enough, we can expect
that when we get concurrent tests to run properly, it will also be not
flaky. This also means we can (in parallel) debug issues related to the
flakiness of individual tests.

See: http://gkecl/152068
Change-Id: Icc3f237b3b5e05ece6024b054ee07f5673c15645