Change bucket path to gs://[PROJECT]/config-sync/
As described in go/cs-release-other-artifacts, the bucket path was
designed to only hold the config-sync-manifest.yaml.  This left no
obvious location for other nomos artifacts (CLI, licenses, etc.).  This
CL changes that path to the more general "config-sync", so that all
nomos artifacts can be deployed to a single path in GCS.

Bug: 161304707
Change-Id: If8a449657748e36b4b882994ebbb94a015129fd8