Remove Makefile.operator
Since the config-sync-manifest.yaml is now committed directly into the
operator (tg/790331), we no longer need the `make operator-manifest`
command, which prepared files to be picked up from the nomos-operator
repository.  This setup expected that the two repositories were both
available under GOPATH, and coupled development in the operator to the
state of the nomso repository.

The Makefile.operator file has essentially nothing left in it, and I've
moved the remaining useful target to a better location.

Bug: 158245995
Change-Id: I46627c076aef436c46d498ca194c2d99845c20a5