Show example cluster creation failures
This is part of debugging why multiple Kind clusters don't work on Prow.

Also log how long creating the kind cluster hangs to help with
debugging.

Increase parallelization to 2 so we can show examples of the job
failing due to this phenomenon.

Increase time to wait for port forward to 20s since it sometimes takes
longer than 10s, and increasing the timeout from 5s to 10s visibly
decreased how often this happened.

Bug: 161793092

Change-Id: I143bba95c9e092735ba1fbfda5580ed57b37127c