Change IFS char in build.sh
One of the cleanup things I did in tg/799359 was to organize the
binaries built in .output/ with the same platform syntax that we
release to GCS with.  That meant:

.output/go/bin/linux-amd64

became

.output/go/bin/linux_amd64 (same as go/acm-downloads#linux)

I forgot to change the build.sh script, and it wasn't able to parse this
new format.  That is fixed in this CL.

Bug: 161304707
Change-Id: I72c45022cf65c5809b2bbcbb5f9f2988c2964121