Increase Go e2e test parallel from 4x to 12x
We've got 4x running smoothly in ~35 minutes, so it feels safe to
increase this to 12x and see if it works. This should reduce runtime to
under 15 minutes.

Runtime is actually being more stable - I misread previous runtimes and
they were actually stable in the range 60-65 minutes. So I feel safe
reducing the timeout from 180m to 120m. As we see more successes it will
be safe to reduce this further.

Bug: 161896218
Change-Id: I921d3dbfdd4c8ffd2b02d89633cf429482254888