Increate Prow test parallel from 12x to 24x
At 12x tests run in ~15-20 minutes, so we're seeing if we can push this
to be even shorter by increasing parallelization further.

The Prow job only has 12vCPUs, and it isn't clear if this directly
translates into threads. So even though we're usually at 1.0/12.0 vCPU
usage, this may slow the job down.

On the other hand, it may speed things up even more, so we're going to
try it.

Bug: 161896218
Change-Id: I6c5f684a9b253d6472fe0d4ea7353d560e8624fc