Increase parallel from 12x to 30x
We're sure this will cause the target to begin to fail, but this means
we'll get a lot of data on _how_ the target fails under adverse
conditions.

I'm setting this value to 30x today so we can make the target more
robust. As-is we only get failures about 15%-20%, of the time, so this
will give us a lot more data.

Alternatively, this will *just work* and tests will be faster.

Change-Id: Ib27915a4c03299270d20ab5d6f1139a83de2dff6