Rewrite flaky bats test in Go
The cluster_resources bats test is super flaky. Rather than debug it,
this rewrites the test using the Go. The flaky test accounts for ~50% of
the Go CI test failures.

This will tell us whether there's a problem with the Go tests, or
whether the test itself was just flaky in bats.

Make waiting for specific types of syncing behavior easy. WaitForSync()
now requires specifying type/name/how to check for sync.
WaitForRepoSync() is for the general "is everything synced" use case.

Change-Id: Ia95964230691403f2521ecb0be658dd93988799f