Updating work queue to contain full objects.
Per the details in go/csmr-watch, we are going to use the work queue in
place of a more standard cache. Therefore it will contain the entire
object to reconcile. We still use the underlying work queue
implementation to enqueue, dequeue, and dedupe GVKNNs. However we also
maintain a map of objects which are keyed off of those GVKNNs. This
object cache favors objects with higher Generation to account for
updates to the object while it is waiting to be reconciled.

Bug: 159821780
Change-Id: I12cff5907c507c81c894ec90e1f4565f4b61e2a4