Convert preserve_fields.bats to Go tests
This is the most recent source of flakiness to Go tests.

Expose the Repository.Root so there is an easy way to write files with
the same YAML printer. This might need refactoring eventually, but would
just make this CL larger.

Add AsSanitizedUnstructured that explicitly removes creationTimestamp
and status, since those can't be in declared configuration. See comment
and linked doc on why we can't automatically do this. b/162260725

Add NT.Kubectl to make calling Kubectl directly easier within tests.
Otherwise it's easy to forget to specify KUBECONFIG, and that's
annoying.

Add HasExactlyAnnotations/HasExactlyLabels for checking the full set of
Annotations/Labels, since we do that in several tests. This shouldn't be
commonplace - only a few tests should actually have to do this, but it's
here to make the transition easier.

Bug: 162329569

Change-Id: I038ddeeaccc0ef4dd22f4f7f23a99fc4d63c2376