Fix objectQueue to override underlying functions.
The previous version had ambiguous behavior because the function
signature did not quite match. For instance, we did not override
AddAfter() because internally it calls Add() which we did override.
However the function paramater type was different so I think AddAfter()
would have called the base version with the matching function signature.
This change makes the overrides explicit and just provides a type-safe
wrapper for use by the watchers and reconcilers.

This also updates the underlying interface to the rate-limiting one
which is what controller runtime uses to throttle requeues during transient
errors.

Bug: 159821780
Change-Id: I2059c442de8c8be87abba8b6951d0a5926ddb165