Translate resource_conditions.bats to Go
This is the last of the three flakiest bats tests.

I think it may be controversial to allow the gatekeeper-specific test
data into these tests, but I think it is worth it. Our paradigm is to
specify as much configuration in the tests *as reasonable*. In this
case, nothing is gained by fully specifying the ConstraintTemplate CRDs
in the tests wheras normally we care about what is in the objects.
Further, I've limited the call to only work for gatekeeper-specific
data. We don't own how this works, so this makes it easier, for example,
to ensure we're actually using the Gatekeeper version of the CRDs.

With this test, we've got all bats tests which have been flaky in the
past 24 hours translated into Go. This should put the pass rate above
the 98% goal we're shooting for.

Bug: 162329569

Change-Id: I22a656fd167fa2e104136499075aab473229b756