Add e2e test that confirms b/162496882
This e2e test confirms that b/162496882 exists. Full explanation in the
bug and in the test that confirms it.

Essentially, we fight with RBAC over the permissions for aggregate
roles. The fix isn't obvious for now, and will probably be very
involved.

It may be fixed with Config Sync v2.

Bug: 162496882
Change-Id: I1c3682d32b84f1d0d3305f94725ad349ed776e07