Add Revert ClusterRole e2e test diagnosing code
The test that ensures we revert a user-modified ClusterRole is flaky.
Let's investigate.

It's possible that either:
1) The ClusterRole is eventually reverted, but we're running too many
simultaneous clusters so running tests is slow, or
2) There is some sort of bug.

We feel confident that if the ClusterRole is not reverted after 10
minutes, it is some sort of bug and it may never be reverted by ACM.
Thus, if the initial test fails, wait another 10 minutes for the revert,
and report how long it took with an error.

This CL won't reduce the rate of errors, but it'll tell us either:
1) How long the ClusterRole is taking to be reverted, or
2) The ClusterRole isn't ever being reverted.

If it is option 1, then we may consider either decreasing test
parallelism or increasing wait time before calling the test a failure.

Change-Id: Ibbd039ba9e6d0db026eb0abbb459986e6bc24db3