Fix test temporary directory error
Glinux got Go updated from 1.14.4 to a development version of 1.15. This
breaks our code that creates temporary directories for our Go e2e tests.

Essentially, the problem is that you can't specify a subdirectory with a
path-separator when creating a temporary directory - any path separators
must be in the base path. This is a change in behavior for
1.14.4 -> 1.15.

Change-Id: I47cc180948fe76448167338d16cfc4223fef2b30