Handle watch.Deleted events from the API server.
This change primarily required work to enable the remediator worker to
understand when an Object in the work queue represented a resource that
had been deleted. We could not just stick a nil Object in the queue
since we need the GVKNN of the Object to properly process it. The first
step was to have the filtered watcher check for Deleted events and make
an explicit call to handle it before passing the Object on to the work
queue. This mirrors what the controller runtime is doing in its
reflector.go file.

I did an initial pass by trying to add a separate map to the objectQueue
of GVKNNs of Objects which had been marked as deleted. This proved
incredibly difficult to maintain in a threadsafe manner while still
respecting the interface dictated by the underlying work queue. In the
end it seemed like there were far too many chances for race conditions.

Will suggested a good approach where we wrap the core.Object in a
package-private struct and test pointers for that wrapped struct. This
is better than my idea of a sentinel annotation map because it does not
mutate the underlying Object at all.

b/159821780

Change-Id: I96c1514fa9034151ff44ad70c79dcf46899158cd