Make b/162828194 failure message clearer
We don't always revert user modifications to declared configuration in a
reasonable time, and this is a bug. Increase the timeout to 20 minutes
to see if maybe it _eventually_ is resolved, or if it is actually just
stuck in that state.

We know we sometimes fight over aggregate rbac permissions. This may be
related to the issue where sometimes we fight with the rbac controller
because of how we serialize/deserialize objects.

It is also possible that this issue is related to us dropping incoming
reconciliation requests under certain (rare) conditions.

Bug: 162828194
Change-Id: I4b851a0c10ac95c171be6ed06ac05483ce5a0d60