Define manager precedence check function.
This function enables a reconciler to determine if it is allowed to
manage/reconcile a resource if it is declared in more than one Git repo.
The filtered watcher can use it now to enable the remediator to ignore
resources it should not remediate.

This CL also includes a lot of refactoring to move the "declared
resources" struct to a new location. It's previous location in the
parser package was not appropriate because the parser actually never
interacts with it.

b/163147932

Change-Id: I47302faf8a842902a024f524d64a1a2755b6d531