Integrate parser into pkg/reconciler/reconciler.go
Now that the CSMR parsers are defined, we can integrate them in the
reconciler code.

This CL does so, and adds the necessary flags and configuration options
to make this work.

Note that FileReader and DiscoveryInterfaceGetter have been left as
options - this will make debugging in e2e situations far easier as we'll
be able to directly control what is being read.

Bug: 163855534
Change-Id: Ia5b979d92ccb3ae4bd5299264d096137a3bd052c