Prevent FileObject nesting
Since ast.FileObject satisfies the core.Object interface, it is easy to
have unknowingly accepted a FileObject as a core.Object, and later try
to parse it into a FileObject.

This CL prevents such cases by checking to see if an object passed to
ParseFileObject is just alrady a FileObject.

Change-Id: I92964f6bb7771fc71ffef7d3a1fc8eb23df0d6e6