Implement manual-test-* make targets
These commands provide a way to manually test configSync during
development.  It includes three commands, each called using `make`:

* manual-test-bootstrap *

This command sets up a working ConfigSync, sans operator.  It is the
same setup done by the e2e tests, and thus is pointed (both in git-creds
and in configmaps) at the test-git-server.  You'll likely need to
override the git-creds and configmaps to fit your usecase.

* manual-test-refresh *

This command reapplies the CS resources, including a newly-built nomos
image.  This is used to update the already bootstrapped CS with your
development changes.  Iterate with this command.

* manual-test-clean *

Remove CS from the cluster.  This uses the test framework's --clean
flag.

Bug: 158858700
Change-Id: Ifc85ae4f4b403a18ed40bbb3cc2844aa08e00cb8