Integrate applier with RepoSync errors.
As part of this work, I noticed that the applier was actually missing a
crucial step when it was listing actual resources. It needs to set the
GVK of the unstructured list to enable the client to know what type to
list.

I also updated structured errors to support conversion to the new
ConfigSyncError type. I also updated the fake testing client to support
label selectors.

b/157609399

Change-Id: Ib437bd5f8617b4178795962dafa0810c000e1e18