Integrate the reconciler manager manifest with the standard layout.
This moves some of the reconciler manager manifest into manifests and splits
out the future manifests/templates/reconciler-manager.yaml from the development
speicfic resources which are now in dev.yaml.  reconciler-manager.yaml is not
being moved to manifests/templates in this chagne since that will require
updating e2e tests to handle filtering the deployment.

Bug: 157609979
Change-Id: I6e1ab4b6f1a9fb29799b4c350e2a97098904f07e