Make nomoserrors warn on missing example errors
nomoserrors now warns if:
- there is an unexpected gap in 1000-level errors
- a known error has no examples

Also make Kptfile type and generators available to other tests.
This required moving some code around to avoid a cyclic import. In
general we want K8s types and code dealing with the types to be in
separate packages for this reason.

Use errors.Is to determine error equality - string comparisons are very
unstable.

Deduplicate string constants that shouldn't have been defined in
multiple locations.

Bug: 165838086

Change-Id: I473bbefe763138a3c9441969249d62042791f598