Explicitly pass Scope when checking diff.Type
Scope applies to the actor - the applier/remediator, not the
declaration. Since any actor has exactly one Scope, we don't need to
rely on parsing the scope written to the objects - we can just use the
one we already have in memory.

Log a message if CanManage detects an invalid non-emptystring manager
annotaiton value. It isn't really an _error_ as we're proceeding
normally even if we discover it, so I'm having it log a warning.

If our filteredWatcher doesn't filter by Namespace before checking
CanManage, then this will cause duplicate warnings for every reconciler.

Also make applier_test more idiomatic.

Bug: 166780454
Change-Id: I86fabc1d2017a379bfa286279a8adb735f12271a