Enable reconcilers to update status.
The CRD definitions were missing the status subresource clause.

The reconcilers also need a rolebinding to do anything on the cluster. I
bound both of them to cluster-admin so that they have authz to write to
the status of their RootSync/RepoSync. The ns reconciler uses a
rolebinding rather than a clusterrolebinding.

Change-Id: Ia1ab092e3e8d31c38f37bf8b41ead6c73373cf53