Make ObjectQueue less confusing.
ObjectQueue implicitly exports its underlying objectQueue, which
implicitly exports its RateLimitingQueue. All of these methods expect or
return interface{}, which does no help to developers trying to
understand what they do.

This resulted in a panic because on retry we pass the core.Object to the
rate limiting interface, when actually we wanted to pass the GVKNN.

Thus, this CL makes it very explicit which level of abstraction a method
is actually using, and removes the ability for callers to pass whatever
they want to the various interfaces.

All methods defined solely by ObjectQueue have been un-exported, and
anything required by other packages is owned by ObjectQueue to avoid any
uncertainty.

On retry, mark the object as dirty so that we don't delete our reference
to it on calling `.Done()` on it.

Also call ShutDown on the internal queue when workers finish.

Bug: 167152774
Change-Id: Ia085dce1fa282ab4511d8e5721f6ac43193d4baf