Don't list CRDs
Rather than also rely on the list of CRDs on the cluster, only check the
API Server for the available types.

I don't believe anything actually needs to list CRDs since:
1) We already use the DiscoveryClient to discover all the resources
already available on the API Server.
2) The existing code intended to parse the CRDClusterConfig, but it
doesn't so the code as written probably has no effect. Essentially this
is a bug in how the old version was supposed to work, but we don't care
to fix it.

Even though (2) is very likely true, I don't want to test that theory in
this CL.

Bug: 167571054

Change-Id: I52c0e5543d16e7405305d577dd59a8691206d74e