Don't panic if cloned git repo is borked
As-is, we panic if we have an error listing files in the git repo. Also
the current code is unintuitive since we actually want to report the
error we got, not immediately bail out.

This CL makes namespace.setSourceStatus safer and clearer with comments,
and structuring to be robust if reading the git directory returned an
error.

Bug: 168018895

Change-Id: Ib3ea6198473bdb67f0578442207a3f795868d85f