Allow managed enabled
In the old code, we added the management=enabled annotations just before
writing the objects to the cluster. In CSMR, we are adding these
annotations before calculating the diff with the cluster.

Thus, diff.Type() shouldn't report an error when it sees the management
annotations since we expect them. In fact, it would be a logic error to
not have the annotations.

This CL changes diff.Type() to expect that resources either have
managed=enabled or managed=disabled. Also change the Applier tests not
to expect Refresh() to ever delete objects.

Bug: 168031204

Change-Id: Iedf73220922e771797b8f97f7429471784c48906