Sync crds in CSMR
Previously, we immediately bailed if any types we wanted were
unavailable on the cluster.

This CL makes it so we track which types we're actually watching (and
can safely sync). Further, it modifies Update() to accept the list of
types we know we can safely watch and sync.

This still leaves two open probleme:
- Our client does not discover new types
- "Invalid value: 0x0: must be specified for an update" errors

Change-Id: I680530da973c0477bdb7c0fad02ba28f83e39785