Conflict resolution errors
If we have stale data from the server, ensure we update our version of
cached resources so we don't endlessly try to send updates computed from
stale versions, or try to update resources which don't exist (and should
be recreated).

Don't wrap Unstructured problems in ObjectParseErrors - the intent there
is for parsing repositories, not objects retrieved from the cluster or
are already in the Applier.

Change-Id: Ie4721a2e310bd01169a365c27a91220fdc1d8377