Force adding example errors
Example errors ensure we know what documentation to update in releases,
so it's essential that we remember to add example errors.

This CL makes it impossible (fine, difficult) to add a new error code
without adding the corresponding examples (and, later, documentation).
The nomoserrors command now complains when error codes are missing
examples, and `make test-unit` runs nomoserrors to ensure this is the
case in presubmits.

Change-Id: I3a14795586f9b9c881f088ca221fcbac308b5206