Enable debug mode in E2E deployments
The main idea here is:

1. Add a --debug flag to the Reconciler main function. When enabled,
reportMisuse() panics instead of just calling glog.Error.

2. Manually modify the ReconcilerManager's ConfigMap so the Reconciler
Deployment has the --debug flag. The Deployment is raw YAML and we don't
want to deal with parsing/unparsing, so this CL just edits the string.
If at any point we aren't able to find what we're looking for or we
aren't able to add the debug flag, this exits the test early.

Note that there aren't tests for this - I have confirmed manually that
this panics when we've got an internal error, and the corresponding E2E
test fails. It isn't clear to me how to have an automated test for this
since as written it requires a bug in our code to trigger a panic. Feel
free to suggest one!

Bug: 169063752

Change-Id: Ide1ddc37f80c5333c6f6bf9cf1d4f7ea03d79b03