Support listing submodule files
This does not include an E2E test - our current E2E setup doesn't
support this. I have, however, verified manually that this works as
intended.

See b/170728505 for why it is infeasible without significant effort to
write an E2E test that could be put into Prow. As-is, it takes ~10
minutes to test manually - just point a cluster at
https://github.com/willbeason/cs-submodule.

Bug: 161824804
Change-Id: I5ce0267f61238a754a53f66518e7bae60ddd07be