Prevent Deployment yaml side-effects with PHONY
willbeason@ noticed that his ConfigSync changes weren't being picked up
by the operator when running `make release-user`.  It turns out that
when tg/851453 removed `rm -rf` from the $(OUTPUT_DIR) target, it
exposed that we weren't using PHONY targets for our deployment
generation.  This was resulting in those targets seeing the existence of
those files and not regenerating.

Change-Id: Ice43a18fed061c390f5708ab4ba0ba0289dbcedd