Always clear lastApplied on error.
We currently cache the last successfully applied commit, and skip
reprocessing if we successuflly applied that commit.  This assumes
that commits hashes will never repeat.  git-sync is OK with force
pushes rewinding history, so it's possible to have commits A -> B -> A
where B is a bad commit that sets errors.  If the user force pushes A
to the tracked branch, the error would not be cleared.

Bug: 169447576
Change-Id: I931dd85f7800f346115e02bf82a4f49a97df6c89