Match GVKs when comparing declared to actual.
The Applier was storing resources by GKNN without checking that the full
GVK matched the declaration. This meant that if Git had multiple versions
of the same kind of resources, we would end up storing an actual
resource with a version that did not match its declaration.

To test and verify, I converted our multiversion BATS test into a
go e2e test. I had to add a version that tests with CRs since core
resources were not triggering this bug.

Bug: 169363387
Change-Id: Ied5e34b5364523098d3a2e9190fa13f0c969fdf9