Introduce "unknown state" tracking for actual.
If we encounter an error when getting/listing a resource from the
cluster, this allows us to explicitly represent that state. Otherwise a
Diff with a missing `actual` is ambiguous: do we know it is missing (and
should be created) or was there an error determining state?

You can see planned usage of it in tg/856471 in applier.go.

Bug: 169633487
Change-Id: I741dae6ee5f9f0149f775c3fdfe6e4b617b1f590