Test synced repos
Working in parallel with Akul on E2E testing CSMR. This CL solves the
problem of "I have CSMR configured, how do I make sure all of my
Namespace repos are synced".

This CL adds functions for automatically testing that Namespace Repos
are synced to the most recent commit.

This does not support:
- checking the sha of multiple branches on the same repository
- syncing a Namespace RepoSync to the Root repository

Neither of these are in the scope of our application; we can test all of
our logic with these limitations.

Change-Id: I23cf973a5e6898d7237f6bd17f77dc6e9fd54ef6