Add conflict CSMR tests
Make RepoSyncClusterRole have permission to edit everything. This keeps
test authors from having to care about RBAC when writing test cases.
Most tests aren't testing how RBAC works, so this keeps tests simpler.
If we want to test specific behavior, we can add that option when we
have tests requiring custom RBAC.

Add check that prevents MonoRepo mode from running if a NamespaceRepo is
specified. Just exit early instead of failing after minutes.

Default ntopts.Control to CentralControl. This is another thing most
tests shouldn't have to care about. The DelegatedControl option, of
course, still exists and can be specified for tests wanting to use that
mode.

Export WaitForRepoSync so that tests can check that the RepoSyncs report
errors at specific commits as desired.

Add two tests that check two patterns for causing conflicts, and both
ways of resolving the conflict.

Fix b/170243322 which was caused by Namespace RepoSyncs not being
checked in Central control mode. Note that Namespace operations still
are not immediate so Retries are not fully eliminated, but reduced to
only waiting for the Namespaces to be deleted.

Bug: 168915540

Change-Id: I13df943043858577c9f941adff7700dda0f2c96a