Use main branch in e2e tests
Git changed the default branch from "master" to "main":
https://github.com/github/renaming

This breaks out E2E tests when running locally as the git CLI creates
the "main" branch, but our tests use the ACM-default "master" branch.

This CL makes us explicitly use the "main" branch in E2E tests.

Also fix a bug in filters.go when the filter is passed the root
directory. Previously we would crash if this happened. In actuality, the
root directory obviously can't be a file and so can be safely excluded.

Bug: 170491291

Change-Id: Id688f23da8480274a20c4227fc7d6841f75223e8