Move RepoSync validation to standardValidation
RepoSync is the single validation step for root repos which is not done
in filesystem.Parser/filesystem.RawParser. Moving it to be done by these
parsers is much easier than adding a layer of abstraction for the sole
purpose of doing a single validation step.

While this could technically cause a false positive if a mono-repo user
defined their own RepoSync type with the same Group we use, this is
vanishingly unlikely.

Context: go/nomos-vet-csmr

Bug: 170979941

Change-Id: I3dfe209f5e1c694a10805bda4bc518b9d47855ee