Add the managed-by annotation to Config Sync
HNC previously respected the `hnc.x-k8s.io/managedBy` annotation to
indicate that another component (e.g. Config Sync) was managing a
namespace, but thanks to HNC's upgrade to v1alpha2, we need to change
this to `hnc.x-k8s.io/managed-by` instead.

This change _adds_ the new annotation but doesn't remove the old one;
that will be done in b/171305869. The reason for this is:

1. We're not 100% certain if v1alpha2 will be ready in time for the next
   release of ACM/CS.
2. I don't want to test all the corner cases of what happens if Config
   Sync is simultaneously removing the old annotation while HNC is still
   using it - I *think* everything will work fine, but it's a lot of
   work to test.
3. Config Sync already adds a bunch of annotations and I don't think the
   (temporary) addition of one more will be a large problem.

Testing: updated the unit and e2e tests and verified that they pass.

Change-Id: I1d25be4c8155b699bda2b00fe54123eb012f9a5f