Allow caching api resources
Users want to vet the scopes of non-core types in CI.

This CL allows users to save the output of `kubectl api-resources` to
their repository for use with nomos vet. Specifically, they must declare
it as "api-resources.txt" in the root of their policy dir.

It is possible for api-resources.txt to conflict with a declared CRD.
For this edge case, the CRD wins as future updates to api-resources.txt
will have the new scope value. This should be rare enough to not worry
about.

Cached API resources are never used when syncing - this is only active
for nomos vet.

Previously DecorateWithClusterConfigs showed an error when talking to a
CSMR repo. Now we do nothing if ClusterConfigs is not defined - by
definition there are no synced CRDs if the type is not defined.

I don't like that this adds an additional method to .Parse(), but I
wasn't able to find a cleaner alternative.

Bug: 172254210
Bug: 155915838

Change-Id: Ie53ee315d84fa23ab37fe55054e7a1d94390454e