Remove `pip install grip` from Dockerfile
Grip is a command-line server application written in Python that uses
the GitHub markdown API to render a local readme file:
https://github.com/joeyespo/grip.

It is used in the `build-buildenv` task in `Makefile.build`. The
`build-cli` task depends on the image built from the `build-buildenv`
task, which builds the nomos binaries. `grip` is not needed in the
process, so this commit removes it to address the security concerns.

b/169712336

Change-Id: I2665ec7fe9c22b0f000264b1cfd39bf28b5fbec0