Don't use upx for darwin
Apparently it is a recurring problem for upx + Go binaries to not work
on Darwin 16 and later.

This is a known recurring issue with upx, that the upx project sees no
real way to fix due to Apple licensing issues.

https://github.com/upx/upx/issues/222

Our best bet is to just not use upx for darwin.

Bug: 171621511
Change-Id: If9f506a1ced97a9741b0eb89b83704eb582f1d11