Move 00-namespace to manifests/test-resources (v2)
This was originally attempted in tg/855840 but was reverted in tg/860461
as the original change broke CI.  The underlying problem was that the
BATS version of the e2e tests was broken, but is not run as part of the
presubmit checks.

As the underlying problem was fixed in a previous CL (tg/875631), we can
now move 00-namespace.yaml into a test-resources folder that is fits its
usage.

Bug: 168827876
Change-Id: I45f6cfd45d7f29a82bd332ef19f45fb350fd5f91