Refactor E2E framework to allow non-Kind clusters
This doesn't actually add any support for non-Kind clusters, it just
makes it possible to do so.

The big change to pay attention to is in new.go, where New() now checks
if the resulting optsStruct has a RESTConfig. If not, it assumes the
test is for Kind. This use pattern will change slightly for allowing
connecting to GKE/etc.; I'd rather wait to see what it looks like than
call that "final".

What will stay the same is that ntopts.Kind() is a new Opt for tests,
which can use it to specify the version of cluster to connect to. This
doesn't change the behavior of the existing Kubernetes version flag -
as-is just means tests could overwrite the behavior, but really this is
used in the future injection point for non-Kind clusters.

Docker-specific logic is mostly moved to its own package - ntopts now
relies on some constants used for Docker logic, so rather than splitting
the constants off the logic now lives in its own package. this is in
line with the desire to split the nomostest package anyway.

The logic for creating Kind clusters is moved to ntopts (to avoid a
circular dependency). The logic is unchanged - just moved around.

Bug: 167659248

Change-Id: I749c489127baa82f0effc99747a152a7d892f43a