Label E2E test objects
A prerequisite for running tests on non-ephemeral clusters is making
sure that clusters are clean before and after tests. We must clean
before as tests aren't guaranteed to exit normally.

This CL adds a TestLabel to all test objects. On startup, the test
cleans the connected cluster before doing anything by checking for the
label in any types in the test's runtime.Schema. All tests use the same
function to generate the runtime.Schema, so they all clean the clusters
consistently. Further, NT now ensures any objects that are created/etc
are registered in the scheme and carry the label. Obviously this isn't
perfect, but it should cover the large majority of cases.

This is the last piece before allowing E2E tests to run on non-ephemeral
clusters.

Notably, this is done invisibly to tests so in the majority of cases
test authors won't need to interact with these labels.

Bug: 167659248

Change-Id: I2f5fc72df7bae9c1d732752f78e93a4ceab7da94