Allow connecting to remote cluster for E2E tests
Add a new test option, RemoteCluster, that tells the test to attach to a
remote cluster and run against it.

Add a new flag, --image-prefix, that says where the test clusters should
find the ConfigSync images. A remote cluster can't use a local docker
registry, so callers must specify a gcr.io location.

Refactor the Clean() function, and only execute it for remote clusters.
Expand the scope of Clean() to also wait for all such objects to be
deleted before terminating. This ensures we don't end up starting a new
test before all Namespaces have been deleted.

Make nt.Wait() into Wait() - this makes it more flexible, and the
wait/retry functionality can be separated into a new package as we want
to break this package up later anyway.

Bug: 167659248

Change-Id: Iafce0efb097822db04fd27bbeb8558c13d199f4b