Implement Go E2E stress tests
I'm requiring the tests to be run manually for now as essentially if
these are run on the wrong cluster they can really mess things up.

In the future we may set up a Prow job that runs these.

For Go E2E tests, conver types from Unstructured to structs so we avoid
dealing with getting/assigning unstructured fields.

Fix setupDelegatedControl() to work for multiple or zero Namespace
repos.

Bug: 167659248

Change-Id: I7392afc3f8d4fbc02c5eb0235a8990fcda4f7456