Fix a side-effects problem and cleanup Makefiles
In building
gs://config-management-release/config-sync/tag/v1.6.0-rc.1/config-sync-manifest.yaml,
Sandy noticed that an extra Deployment ended up in there: syncer

This was because the make assumed that the .output directory was clean,
which is hard to actually insure.  This CL adds a clean-deployments-dir
dependency step to wipe out the contents of .output/deployment and
prevent this problem.  It also cleans up some variable and target names
to better represent their actual usages.

I thought about doing a clean wipe of OUTPUT_DIR, but it turns out that
this is hard to sequence easily.  Many targets list OUTPUT_DIR as a
dependency, and thus we can't just go about deleting it over and over
again.  Like it or not, .output existing between command runs is an
assumption of this infrastructure.  The solution would be a temporary
filesystem in docker.

Bug: 173249258
Change-Id: Ic342e0df78e84b0bdfdb04d536a9d5fc92327e54