Move validation from noClusterScopeValidator to scopeValidator
Now we are passing `CLUSTER_NAME` to the namespace reconciler, we need
to validate if the namespace repo contains cluster-scoped resources,
like Cluster or ClusterSelectors, in case the namespace-scoped resources
are filtered out accidentally by the invalid ClusterSelectors.

The noClusterScopeValidator validates the object scope after the
raw_parser parses the file objects. We would like to make it fail
earlier in the process of raw parsing.

b/170982143

Change-Id: I8d4e5d7f271623d638cc894525af7610c5a017bc