Add Anthos Validator to built-in scopes
A use case ConfigSync does _not_ support is users declaring resources
which they want neither synced nor validated. This causes problems for
the anthos Validator type, as it will only ever be marked managed:
disabled, and shouldn't be synced to clusters.

We have no mechanism for specifying this for an entire type. If a user
declares the resources (which they have to, inside policydir) and even
sets it managed: disabled, we will fail if the resource's CRD is not
defined or in the cluster.

This solution just adds the Anthos Validator type to the list of known
kinds. This sidesteps the issue so a properly configured (i.e.
managed: disabled) Validator will cause no problems.

There are edge cases, but they should be rare enough to not care about
(e.g. invalid metadata.name, forbidden annotations). If this is a
problem, we will need to design a solution.

In testing this change I found that nomos vet doesn't natively have
ConfigSync types registered in the parser, so if used this causes a
KNV1006 when not connected to a cluster if the repository declared
configsync.gke.io types.

Bug: 161812252

Change-Id: Iaa5f2a6e9ebf92a2dd2c9d180b025f4baeb3ed9d