Upgrade nomos dependencies to v0.19.4
This upgrades our libraries from the inconsistent, unsupported
combination of individual commits that Google3 uses to the most recent
stable set of libraries.

Why now? Because we don't have much changing now that 1.6.0 is released,
and waiting until we *have* to upgrade creates a lot of unnecessary
pain.

This almost completely eliminates our replace directive - the only one
I've left is to make it explicit which version of code-generator we're
using. As-is the directive does nothing, but is to make it easier for
maintainers to know/coordinate which version to use.

The non-vendor/non-generated-code changes are almost exclusively adding
context.Context where the new libraries now require it.

In the upgrade, the Kubernetes libraries officially dropped support for
"preserveUnknownFields: true" - all CRDs are implicitly structural. I've
removed the test for structural schemas for v1 CRDs as the test isn't
actually testing anything in a 1.16+ world, and the field's presence is
an error in 1.19+.

Additionally, I've removed the YAMLs defining invalid Anvil CRDs. They
aren't exercised by any existing E2E tests. It isn't clear what paths
they were originally used for testing.

Bug: 173730710

Change-Id: Ie6f840bef6692500178dae836fdfbb506116bf0c